refactor: split indexing services
